+++ emp-manager/src/main/java/com/example/empmanager/EmployeeResource.java
@@ -1,0 +1,50 @@
+package com.example.empmanager;
+
+import com.example.empmanager.model.Employee;
+import com.example.empmanager.service.EmployeeService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/employee")
+public class EmployeeResource {
+
+    private final EmployeeService employeeService;
+
+    public EmployeeResource(EmployeeService employeeService) {
+        this.employeeService = employeeService;
+    }
+
+    @GetMapping("/all")
+    public ResponseEntity<List<Employee>> getAllEmployees() {
+        final List<Employee> all = employeeService.findAll();
+        return new ResponseEntity<>(all, HttpStatus.OK);
+    }
+
+    @GetMapping("/find/{id}")
+    public ResponseEntity<Employee> getEmployee(@PathVariable("id") Long employeeId) {
+        Employee employee = employeeService.findById(employeeId);
+        return new ResponseEntity<>(employee, HttpStatus.OK);
+    }
+
+    @PostMapping("/add")
+    public ResponseEntity<Employee> addEmployee(@RequestBody Employee employee) {
+        final Employee saved = employeeService.addEmployee(employee);
+        return new ResponseEntity<>(saved, HttpStatus.CREATED);
+    }
+
+    @PutMapping("/update")
+    public ResponseEntity<Employee> updateEmployee(@RequestBody Employee employee) {
+        final Employee saved = employeeService.updateEmployee(employee);
+        return new ResponseEntity<>(saved, HttpStatus.OK);
+    }
+
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<Void> deleteEmployee(@PathVariable("id") Long employeeId) {
+        employeeService.deleteEmployee(employeeId);
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+}