+++ emp-manager/History.md
@@ -4,3 +4,5 @@ History of Implementation
 1. Add Model class Employee.
 2. Database configuration.
 3. Add Employee Repository and Service.
+4. Add bootstrap data.
+5. Add REST Controller to expose API.